Add new config.json file with LogLevel setting and update ServiceOverrides.cs to use SharedServices.Locator for IFileService
Introduces a new config.json file in the Resources folder with the "LogLevel" setting. The ServiceOverrides.cs file is updated to use SharedServices.Locator for the IFileService instead of the previous fallback implementation. This change enables easier configuration of logging levels without requiring code modifications. The Log.cs file in the Runtime/Log folder is updated to utilize the ServiceLocator for IFileService as well.